requirements.txt
streamlit==1.28.0
pandas==2.0.3
plotly==5.15.0
requests==2.31.0
playwright==1.40.0
python-dotenv==1.0.0
gspread==6.0.0
google-auth==2.25.0
gspread-formatting==1.1.2
chromadb>=0.4.0
